+++ sniffer/__main__.py
@@ -4,29 +4,31 @@
     that traffic.
 """
 
+import sys
 import argparse
-from argparse import RawDescriptionHelpFormatter
 from .__version__ import __version__
 from .network.engine import SnifferEngine
+from .utils.constants import EXIT_SUCCESS
 from .utils.decorators import require_root
 from .utils.constants import PARSER_IGNORE
+from .parser.options import list_interfaces
 from .parser.textutils import BANNER, EPILOG
-from .parser.ColoredArgumentParser import ColoredArgumentParser
+from .parser.custom import ColoredArgumentParser
 
 
 def create_parser():
     parser = ColoredArgumentParser(
+        prog='sniffer' if sys.argv[0] == '-m' else sys.argv[0],
         description=f'{BANNER}\n{__doc__}',
-        formatter_class=RawDescriptionHelpFormatter,
+        formatter_class=lambda prog: argparse.RawDescriptionHelpFormatter(
+            prog,
+            max_help_position=40
+        ),
         add_help=False,
         epilog=EPILOG
     )
 
     positional_args = parser.add_argument_group('POSITIONAL ARGUMENTS')
-    positional_args.add_argument(
-        'a',
-        help='[blue]a first argument[/blue]'
-    )
 
     optional_args = parser.add_argument_group('OPTIONAL ARGUMENTS')
     optional_args.add_argument(
@@ -37,17 +39,18 @@ def create_parser():
              '[/blue]'
     )
     optional_args.add_argument(
-        '-n',
-        '--no-color',
-        help='[blue]Disable output coloring.[/blue]',
-        action='version'
-    )
-    optional_args.add_argument(
         '-i',
         '--interface',
         help='[blue]The network interface to be used. Default: any[/blue]',
         default='any',
         dest='interface'
+    )
+    optional_args.add_argument(
+        '-l',
+        '--list-interfaces',
+        help='[blue]List all network interfaces present in the system[/blue]',
+        dest='list_interfaces',
+        action='store_true'
     )
     optional_args.add_argument(
         '-v',
@@ -74,6 +77,20 @@ def create_parser():
         nargs='?'   # Bypass the required value for positional arguments
     )
 
+    # By parsing all the arguments until here, we are able to bypass the need
+    # to fill the positional arguments
+    parsed, _ = parser.parse_known_args()
+
+    if parsed.list_interfaces:
+        list_interfaces()
+        exit(EXIT_SUCCESS)
+
+    # This will be parsed after checking all optional arguments
+    positional_args.add_argument(
+        'a',
+        help='[blue]a first argument[/blue]'
+    )
+
     return parser.parse_args()
 
 
@@ -85,8 +102,8 @@ def sniff(interface):
 
 def main():
     args = create_parser()
-    print(f'Using {args.interface}')
 
+    print(f'Using {args.interface}')
     sniff(args.interface)