+++ CHANGELOG.md
@@ -14,6 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Create `decorators.py` file inside `utils` package that will contain, you've
 guessed it: decorators. Also, created the first decorator, `require_root` that
 will exit if the user running the script isn't **root**
+- Implement `-l/--list-interfaces` option, which will list all interfaces
+present in the system together with their MAC address + IPv4/6 address
+
+### Changed
+- Update Sniffer's description in `README.md` file
 
 ### Fixed
 - Fix Layer 4 header size being (wrongly) too broad